requirements.txt
disnake
squarecloud-api
mercadopago
requests
python-dotenv